// src/types/recruitment.ts
export interface RecruitmentType {
    id: string;
    recruitmentTitle: string;
    content: string;
    position: number;
    experience: string;
    quantity: string;
    createdAt: string;
    updatedAt: string;
    expirationDate: string;
    status: boolean;
    typesOfWork: string;
    salaryRange: string;
    metaDescription: string | null;
    urlPath: string;
  };